Cherrypick/2680 (#2688)
* prover v3.0.0-RC3-fork.6

* prover v3.0.0

* update prover to v3.0.0-RC1

* update prover to v0.3.1-RC1

* update tests to forkID6

* update tests to forkID6

* update prover image

* udpate test

* Changing the forced batch tests to use genesisBlockNumber from the test.genesis.config.json.


* Fixing forced batch tests to be symlinks and fixing usage of SetForkID.


* Fixing forced batch tests to be symlinks and fixing usage of SetForkID.


* update prover image to v3.0.0-RC2

---------
