+++ ci/e2e-group11/shared.go
@@ -1,0 +1,1 @@
+../../test/e2e/shared.go
