+++ docker-compose.yml
@@ -107,7 +107,7 @@ services:
   zkevm-prover:
     container_name: zkevm-prover
     restart: unless-stopped
-    image: hermeznetwork/zkevm-prover:v2.2.3
+    image: hermeznetwork/zkevm-prover:v3.0.0-RC2
     depends_on:
       zkevm-state-db:
         condition: service_healthy
