+++ test/docker-compose.yml
@@ -320,7 +320,7 @@ services:
 
   zkevm-mock-l1-network:
     container_name: zkevm-mock-l1-network
-    image: hermeznetwork/geth-zkevm-contracts:v2.0.0-RC1-fork.5-geth1.12.0
+    image: hermeznetwork/geth-zkevm-contracts:v2.0.0-fork.6-geth1.12.0
     ports:
       - 8545:8545
       - 8546:8546
@@ -348,7 +348,7 @@ services:
 
   zkevm-prover:
     container_name: zkevm-prover
-    image: hermeznetwork/zkevm-prover:v2.2.3
+    image: hermeznetwork/zkevm-prover:v3.0.0-RC2
     ports:
       - 50061:50061 # MT
       - 50071:50071 # Executor
@@ -422,7 +422,7 @@ services:
 
   zkevm-permissionless-prover:
     container_name: zkevm-permissionless-prover
-    image: hermeznetwork/zkevm-prover:v2.2.3
+    image: hermeznetwork/zkevm-prover:v3.0.0-RC2
     ports:
       # - 50058:50058 # Prover
       - 50059:50052 # Mock prover
