+++ test/e2e/forced_batches_vector_group2_test.go
@@ -9,32 +9,28 @@ import (
 	"testing"
 	"time"
 
-	"github.com/0xPolygonHermez/zkevm-node/etherman/smartcontracts/polygonzkevm"
+	"github.com/0xPolygonHermez/zkevm-node/config"
 	"github.com/0xPolygonHermez/zkevm-node/hex"
 	"github.com/0xPolygonHermez/zkevm-node/log"
 	"github.com/0xPolygonHermez/zkevm-node/state"
-	"github.com/0xPolygonHermez/zkevm-node/test/constants"
 	"github.com/0xPolygonHermez/zkevm-node/test/operations"
 	"github.com/0xPolygonHermez/zkevm-node/test/vectors"
-	"github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/stretchr/testify/require"
 )
 
-const (
-	forkID = 5
-)
-
-func TestForcedBatchesVectorFiles(t *testing.T) {
+func TestForcedBatchesVectorFilesGroup2(t *testing.T) {
 
 	if testing.Short() {
 		t.Skip()
 	}
-	vectorFilesDir := "./../vectors/src/state-transition/forced-tx"
+	vectorFilesDir := "./../vectors/src/state-transition/forced-tx/group2"
 	ctx := context.Background()
-	err := filepath.Walk(vectorFilesDir, func(path string, info os.FileInfo, err error) error {
+	genesisFileAsStr, err := config.LoadGenesisFileAsString("../../test/config/test.genesis.config.json")
+	require.NoError(t, err)
+	genesisConfig, err := config.LoadGenesisFromJSONString(genesisFileAsStr)
+	require.NoError(t, err)
+	err = filepath.Walk(vectorFilesDir, func(path string, info os.FileInfo, err error) error {
 		if err != nil {
 			return err
 		}
@@ -63,7 +59,8 @@ func TestForcedBatchesVectorFiles(t *testing.T) {
 				log.Info("# Setting Genesis #")
 				log.Info("###################")
 				genesisActions := vectors.GenerateGenesisActions(testCase.Genesis)
-				require.NoError(t, opsman.SetGenesis(genesisActions))
+				require.NoError(t, opsman.SetGenesis(genesisConfig.Genesis.GenesisBlockNum, genesisActions))
+				require.NoError(t, opsman.SetForkID(genesisConfig.Genesis.GenesisBlockNum, forkID6))
 				require.NoError(t, opsman.Setup())
 
 				// Check initial root
@@ -75,7 +72,7 @@ func TestForcedBatchesVectorFiles(t *testing.T) {
 				require.Equal(t, testCase.ExpectedOldStateRoot, actualOldStateRoot.Hex())
 				decodedData, err := hex.DecodeHex(testCase.BatchL2Data)
 				require.NoError(t, err)
-				_, txBytes, _, err := state.DecodeTxs(decodedData, forkID)
+				_, txBytes, _, err := state.DecodeTxs(decodedData, forkID6)
 				forcedBatch, err := sendForcedBatchForVector(t, txBytes, opsman)
 				require.NoError(t, err)
 				actualNewStateRoot := forcedBatch.StateRoot
@@ -134,99 +131,4 @@ func TestForcedBatchesVectorFiles(t *testing.T) {
 		return nil
 	})
 	require.NoError(t, err)
-}
-
-func sendForcedBatchForVector(t *testing.T, txs []byte, opsman *operations.Manager) (*state.Batch, error) {
-	ctx := context.Background()
-	st := opsman.State()
-	// Connect to ethereum node
-	ethClient, err := ethclient.Dial(operations.DefaultL1NetworkURL)
-	require.NoError(t, err)
-
-	// Create smc client
-	zkEvmAddr := common.HexToAddress(operations.DefaultL1ZkEVMSmartContract)
-	zkEvm, err := polygonzkevm.NewPolygonzkevm(zkEvmAddr, ethClient)
-	require.NoError(t, err)
-
-	auth, err := operations.GetAuth(operations.DefaultSequencerPrivateKey, operations.DefaultL1ChainID)
-	require.NoError(t, err)
-
-	log.Info("Using address: ", auth.From)
-	num, err := zkEvm.LastForceBatch(&bind.CallOpts{Pending: false})
-	require.NoError(t, err)
-	log.Info("Number of forceBatches in the smc: ", num)
-
-	// Get tip
-	tip, err := zkEvm.GetForcedBatchFee(&bind.CallOpts{Pending: false})
-	require.NoError(t, err)
-
-	disallowed, err := zkEvm.IsForcedBatchDisallowed(&bind.CallOpts{Pending: false})
-	require.NoError(t, err)
-	if disallowed {
-		tx, err := zkEvm.ActivateForceBatches(auth)
-		require.NoError(t, err)
-		err = operations.WaitTxToBeMined(ctx, ethClient, tx, operations.DefaultTimeoutTxToBeMined)
-		require.NoError(t, err)
-	}
-
-	// Send forceBatch
-	tx, err := zkEvm.ForceBatch(auth, txs, tip)
-	require.NoError(t, err)
-
-	log.Info("Forced Batch Submit to L1 TxHash: ", tx.Hash())
-	time.Sleep(1 * time.Second)
-
-	err = operations.WaitTxToBeMined(ctx, ethClient, tx, operations.DefaultTimeoutTxToBeMined)
-	require.NoError(t, err)
-
-	currentBlock, err := ethClient.BlockByNumber(ctx, nil)
-	require.NoError(t, err)
-	log.Debug("currentBlock.Time(): ", currentBlock.Time())
-
-	query := ethereum.FilterQuery{
-		FromBlock: currentBlock.Number(),
-		Addresses: []common.Address{zkEvmAddr},
-	}
-	logs, err := ethClient.FilterLogs(ctx, query)
-	require.NoError(t, err)
-
-	var forcedBatch *state.Batch
-	for _, vLog := range logs {
-		if vLog.Topics[0] != constants.ForcedBatchSignatureHash {
-			logs, err = ethClient.FilterLogs(ctx, query)
-			require.NoError(t, err)
-			continue
-		}
-		fb, err := zkEvm.ParseForceBatch(vLog)
-		if err != nil {
-			log.Errorf("failed to parse force batch log event, err: ", err)
-		}
-		log.Debugf("log decoded: %+v", fb)
-		ger := fb.LastGlobalExitRoot
-		log.Info("GlobalExitRoot: ", ger)
-		log.Info("Transactions: ", common.Bytes2Hex(fb.Transactions))
-		fullBlock, err := ethClient.BlockByHash(ctx, vLog.BlockHash)
-		if err != nil {
-			log.Errorf("error getting hashParent. BlockNumber: %d. Error: %v", vLog.BlockNumber, err)
-			return nil, err
-		}
-		log.Info("MinForcedTimestamp: ", fullBlock.Time())
-		forcedBatch, err = st.GetBatchByForcedBatchNum(ctx, fb.ForceBatchNum, nil)
-		for err == state.ErrStateNotSynchronized {
-			time.Sleep(1 * time.Second)
-			forcedBatch, err = st.GetBatchByForcedBatchNum(ctx, fb.ForceBatchNum, nil)
-		}
-		require.NoError(t, err)
-		require.NotNil(t, forcedBatch)
-
-		log.Info("Waiting Forced Batch to be virtualized ...")
-		err = operations.WaitBatchToBeVirtualized(forcedBatch.BatchNumber, 4*time.Minute, st)
-		require.NoError(t, err)
-
-		log.Info("Waiting Forced Batch to be consolidated ...")
-		err = operations.WaitBatchToBeConsolidated(forcedBatch.BatchNumber, 4*time.Minute, st)
-		require.NoError(t, err)
-	}
-
-	return forcedBatch, nil
 }
